+++ src/main/java/com/example/credit_service_project/DTO/operationDTO/PreliminaryCalculationRequest.java
@@ -10,7 +10,7 @@ import java.util.UUID;
 public class PreliminaryCalculationRequest {
     UUID accountID;
     String accountNumber;
-    BigInteger productID;
+    Long productID;
     Integer monthTerm;
     BigDecimal interestRate;
 }
