+++ src/main/java/com/example/credit_service_project/services/product/ProductDeleteService.java
@@ -7,8 +7,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.math.BigInteger;
-
 @Service
 @RequiredArgsConstructor
 @Slf4j
@@ -16,7 +14,7 @@ public class ProductDeleteService {
     private final ProductSearchService searchProductService;
     private final ProductUtil util;
 
-    public ProductResponseDTO deleteProduct(BigInteger id) {
+    public ProductResponseDTO deleteProduct(Long id) {
         Product product = searchProductService.findById(id);
         log.info("Delete product with ID: {}", id);
         return util.toResponse(product);
